Override build  Authentication method (/login endpoint)
