+++ src/main/java/edu/kpi/testcourse/exception/user/UserNotFoundException.java
@@ -1,0 +1,6 @@
+package edu.kpi.testcourse.exception.user;
+
+/**
+ * Throw this exception if a user is not in the database.
+ */
+public class UserNotFoundException extends RuntimeException {}
